requirements.txt
azure-cognitiveservices-speech==1.25.0
openai==0.27.0
pandas==1.5.3
numpy==1.24.2
pyttsx3==2.90
gTTS==2.3.1
python-dotenv==0.21.1
matplotlib==3.7.1
requests==2.28.2
scipy==1.10.1
scikit-learn==1.2.1